Hot-fix Bug-7-Log-Reg
- Rename controller (user.js to the auth.js)
- Join two services (auth.js and registrationUser.js to the one auth.js)
- Create the folder "helpers"
- Rename user.validation.js to the auth.js
- Rename keys to the schema
- Separate express error handler to another folder
- Insert token into registration
