+++ api/helpers/index.js
@@ -1,0 +1,7 @@
+const hashHelpers = require('./hashHelpers');
+const jwtHelpers = require('./jwtHelpers');
+
+module.exports = {
+  hashHelpers,
+  jwtHelpers,
+};
