+++ api/routes/user.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const controllers = require('../controllers/user');
+const controllers = require('../controllers/auth');
 
 router.post('/login', controllers.login);
