--- api/services/error.handler.js
@@ -1,3 +1,0 @@
-module.exports = (error, options) => {
-  options && options.show && console.log(error);
-};
