+++ index.js
@@ -2,7 +2,9 @@ require('dotenv').config();
 
 const express = require('express');
 const bodyParser = require('body-parser');
+
 const { user } = require('./api/routes');
+const { errorHandlerMiddleware } = require('./api/middlewares');
 
 const app = express();
 
@@ -17,13 +19,7 @@ app.use('/api/users/', user);
  * TODO:  Rewrite processing of errors from login and registration routers
  */
 
-app.use(function(error, req, res, next) {
-  if (error.message) {
-    res.status(400).json({ message: error.message });
-  } else {
-    res.status(400).send({ status: 400, message: error.toString() });
-  }
-});
+app.use(errorHandlerMiddleware);
 
 app.listen(PORT, function() {
   console.log(`Listening on port ${PORT}`);
