+++ package.json
@@ -12,7 +12,6 @@
   "license": "ISC",
   "dependencies": {
     "@hapi/joi": "^15.1.0",
-    "bcrypt": "^3.0.6",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "dotenv": "^8.0.0",
